feat: cuda tracegen + tests for divrem (#108)
* Cuda tracegen + tests for Rv32DivRem

* review comments

+++ tracegen-gpu/src/extensions/rv32im/hintstore.rs
@@ -101,10 +101,10 @@ impl DeviceChip<SC, GpuBackend> for Rv32HintStoreChipGpu {
                 &d_records,
                 offsets.len() as u32,
                 &d_record_offsets,
-                self.pointer_max_bits,
+                self.pointer_max_bits as u32,
                 &self.range_checker.count,
                 &self.bitwise_lookup.count,
-                RV32_CELL_BITS,
+                RV32_CELL_BITS as u32,
             )
             .unwrap();
         }
